refactor 1 for 1

+++ server/internal/services/settings/model/icon.go
@@ -1,7 +1,7 @@
 package model
 
 type Icon struct {
-	ID   int    `json:"id" db:"id"`     // ID иконки
+	ID   int64  `json:"id" db:"id"`     // ID иконки
 	Name string `json:"name" db:"name"` // Название иконки
 	Url  string `json:"url" db:"img"`   // URL иконки
 }
